+++ src/components/shared/cart/MiniCart.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { addCartItem, clearCart, fetchCart, removeCartItem } from "@/store/cart"
+import { addCartItem, clearCart, fetchCart, removeCartItem, selectCart } from "@/store/cart"
 import { useAppSelector } from "@/libs/redux/hooks"
 import { SummaryField } from "./SummaryField"
 import { AppDispatch } from "@/store/store"
@@ -8,10 +8,8 @@ import { CartItemCard } from "./CartItem"
 import { useEffect, useRef } from "react"
 import { useDispatch } from "react-redux"
 import { useRouter } from "next/navigation"
-import { fetchTaxes, setCheckoutStatus } from "@/store/checkout"
+import { fetchTaxes, setCheckoutStatus, selectOrderStatus } from "@/store/checkout"
 import { handleHttpErrors } from "@/utils/helpers"
-import { selectOrderStatus } from "@/store/checkout/checkoutSlice"
-import { selectCart } from "@/store/cart/cartSlice"
 import { PrimaryButton } from "../buttons/PrimaryButton"
 import { selectUserStatus } from "@/store/user/userSlice"
