+++ src/store/cart/index.ts
@@ -1,2 +1,9 @@
 export { addCartItem, fetchCart, clearCart, removeCartItem } from "./thunks"
-export { cartReducer } from "./cartSlice"
+export {
+    cartReducer,
+    selectCart,
+    selectCartItems,
+    selectCartStatus,
+    selectCartItemsLength,
+    selectBusyProductsLength
+} from "./cartSlice"
