+++ src/store/checkout/index.ts
@@ -1,2 +1,7 @@
-export { setCheckoutStatus, checkoutReducer } from "./checkoutSlice"
+export {
+    setCheckoutStatus,
+    checkoutReducer,
+    selectOrderStatus,
+    selectTaxesStatus
+} from "./checkoutSlice"
 export { createOrder, fetchTaxes } from "./thunks"
